chore(core): replace vault detail screen with 2.0
