+++ mobile-app/app/screens/AppNavigator/screens/Loans/VaultDetail/VaultDetailScreen.tsx
@@ -1,108 +1,141 @@
-import { View } from "@components";
-import {
-  ThemedIcon,
-  ThemedScrollView,
-  ThemedText,
-  ThemedView,
-} from "@components/themed";
+import { ThemedScrollViewV2 } from "@components/themed";
 import { StackScreenProps } from "@react-navigation/stack";
 import { tailwind } from "@tailwind";
-import { translate } from "@translations";
 import { useEffect, useState } from "react";
-import { TouchableOpacity } from "react-native";
-import { fetchCollateralTokens, LoanVault, vaultsSelector } from "@store/loans";
-import { useSelector } from "react-redux";
-import { RootState } from "@store";
-import { LoanVaultState } from "@defichain/whale-api-client/dist/api/loan";
-import BigNumber from "bignumber.js";
-import { useDeFiScanContext } from "@shared-contexts/DeFiScanContext";
+import { Platform, View } from "react-native";
 import { openURL } from "@api/linking";
 import {
-  useVaultStatus,
-  VaultStatusTag,
-} from "@screens/AppNavigator/screens/Loans/components/VaultStatusTag";
-import { CollateralizationRatioDisplay } from "@screens/AppNavigator/screens/Loans/components/CollateralizationRatioDisplay";
+  fetchCollateralTokens,
+  loanTokensSelector,
+  LoanVault,
+  vaultsSelector,
+} from "@store/loans";
+import { WalletAlert } from "@components/WalletAlert";
+import { useSelector } from "react-redux";
+import { RootState } from "@store";
+import {
+  LoanToken,
+  LoanVaultActive,
+  LoanVaultState,
+  LoanVaultTokenAmount,
+} from "@defichain/whale-api-client/dist/api/loan";
+import BigNumber from "bignumber.js";
+import { useVaultStatus } from "@screens/AppNavigator/screens/Loans/components/VaultStatusTag";
 import { useNextCollateralizationRatio } from "@screens/AppNavigator/screens/Loans/hooks/NextCollateralizationRatio";
 import { useLoanOperations } from "@screens/AppNavigator/screens/Loans/hooks/LoanOperations";
-import { VaultStatus } from "@screens/AppNavigator/screens/Loans/VaultStatusTypes";
-import { getPrecisedTokenValue } from "@screens/AppNavigator/screens/Auctions/helpers/precision-token-value";
 import { useAppDispatch } from "@hooks/useAppDispatch";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { useIsFocused } from "@react-navigation/native";
-import { VaultSectionTextRow } from "../components/VaultSectionTextRow";
-import { VaultDetailTabSection } from "./components/VaultDetailTabSection";
-import { ScrollableButton, ScrollButton } from "../components/ScrollableButton";
+import { VaultDetailStatus } from "@screens/AppNavigator/screens/Loans/VaultDetail/components/VaultDetailStatus";
+import { VaultActionButtons } from "@screens/AppNavigator/screens/Loans/VaultDetail/components/VaultActionButtons";
+import {
+  BottomSheetWebWithNavV2,
+  BottomSheetWithNavV2,
+} from "@components/BottomSheetWithNavV2";
+import { useBottomSheet } from "@hooks/useBottomSheet";
+import { BottomSheetLoanTokensList } from "@screens/AppNavigator/screens/Loans/components/BottomSheetLoanTokensList";
+import { useThemeContext } from "@shared-contexts/ThemeProvider";
+import { CloseVaultButton } from "@screens/AppNavigator/screens/Loans/VaultDetail/components/CloseVaultButton";
+import { VaultDetailSummary } from "@screens/AppNavigator/screens/Loans/VaultDetail/components/VaultDetailSummary";
+import { useMaxLoan } from "@screens/AppNavigator/screens/Loans/hooks/MaxLoanAmount";
+import { BottomSheetPayBackList } from "@screens/AppNavigator/screens/Loans/components/BottomSheetPayBackList";
+import { useCollateralTokenList } from "@screens/AppNavigator/screens/Loans/hooks/CollateralTokenList";
+import { CollateralItem } from "@screens/AppNavigator/screens/Loans/screens/EditCollateralScreen";
+import {
+  BottomSheetToken,
+  BottomSheetTokenList,
+  TokenType,
+} from "@components/BottomSheetTokenList";
+import { BottomSheetTokenListHeader } from "@components/BottomSheetTokenListHeader";
+import { translate } from "@translations";
+import { useDeFiScanContext } from "@shared-contexts/DeFiScanContext";
+import { VaultDetailLoansRow } from "./components/VaultDetailLoansRow";
+import { VaultDetailCollateralsRow } from "./components/VaultDetailCollateralsRow";
 import { LoanParamList } from "../LoansNavigator";
 
 type Props = StackScreenProps<LoanParamList, "VaultDetailScreen">;
 
 export function VaultDetailScreen({ route, navigation }: Props): JSX.Element {
-  const { vaultId, tab } = route.params;
+  const { isLight } = useThemeContext();
+  const { vaultId } = route.params;
   const [vault, setVault] = useState<LoanVault>();
-  const vaults = useSelector((state: RootState) => vaultsSelector(state.loans));
   const canUseOperations = useLoanOperations(vault?.state);
   const dispatch = useAppDispatch();
   const client = useWhaleApiClient();
   const blockCount = useSelector((state: RootState) => state.block.count);
   const isFocused = useIsFocused();
-  const vaultActionButtons: ScrollButton[] = [
-    {
-      label: "EDIT COLLATERAL",
-      disabled: !canUseOperations,
-      handleOnPress: () => {
-        if (vault === undefined) {
-          return;
-        }
+  const { collateralTokens } = useCollateralTokenList();
+  const { getVaultsUrl } = useDeFiScanContext();
 
-        navigation.navigate({
-          name: "EditCollateralScreen",
-          params: {
-            vaultId: vault.vaultId,
-          },
-          merge: true,
-        });
-      },
-      testID: "vault_detail_edit_collateral",
-    },
-    {
-      label: "EDIT LOAN SCHEME",
-      disabled: !canUseOperations || vault?.state === LoanVaultState.FROZEN,
-      handleOnPress: () => {
-        if (vault === undefined) {
-          return;
-        }
+  const vaults = useSelector((state: RootState) => vaultsSelector(state.loans));
+  const loanTokens = useSelector((state: RootState) =>
+    loanTokensSelector(state.loans)
+  );
 
-        navigation.navigate({
-          name: "EditLoanSchemeScreen",
-          params: {
-            vaultId: vault.vaultId,
-          },
-          merge: true,
-        });
-      },
-      testID: "vault_detail_edit_loan_scheme",
-    },
-    {
-      label: "CLOSE VAULT",
-      disabled: !(
-        vault?.state === LoanVaultState.ACTIVE && vault.loanValue === "0"
+  const inLiquidation = vault?.state === LoanVaultState.IN_LIQUIDATION;
+  const totalLoanAmount = new BigNumber(
+    inLiquidation || vault === undefined ? 0 : vault?.loanValue
+  );
+  const totalCollateralValue = new BigNumber(
+    inLiquidation || vault === undefined ? 0 : vault?.collateralValue
+  );
+  const minColRatio = new BigNumber(
+    vault === undefined ? 0 : vault?.loanScheme.minColRatio
+  );
+  const vaultState = useVaultStatus(
+    vault?.state,
+    new BigNumber(
+      inLiquidation || vault === undefined ? 0 : vault?.informativeRatio
+    ),
+    minColRatio,
+    totalLoanAmount,
+    totalCollateralValue
+  );
+  const collateralAmounts =
+    inLiquidation || vault === undefined ? [] : vault.collateralAmounts;
+  const loanAmounts =
+    inLiquidation || vault === undefined ? [] : vault.loanAmounts;
+  const nextCollateralizationRatio = useNextCollateralizationRatio(
+    collateralAmounts,
+    loanAmounts
+  );
+  const maxLoanAmount = useMaxLoan({
+    totalCollateralValue: totalCollateralValue,
+    collateralAmounts: collateralAmounts,
+    existingLoanValue: totalLoanAmount,
+    minColRatio: minColRatio,
+  });
+
+  const [snapPoints, setSnapPoints] = useState({
+    ios: ["75%"],
+    android: ["70%"],
+  });
+  const getBottomSheetHeader = (label: string) => {
+    return {
+      headerStatusBarHeight: 2,
+      headerTitle: "",
+      headerBackTitleVisible: false,
+      headerStyle: tailwind("rounded-t-xl-v2 border-b-0", {
+        "bg-mono-light-v2-100": isLight,
+        "bg-mono-dark-v2-100": !isLight,
+      }),
+      header: () => (
+        <BottomSheetTokenListHeader
+          headerLabel={label}
+          onCloseButtonPress={dismissModal}
+        />
       ),
-      handleOnPress: () => {
-        if (vault === undefined) {
-          return;
-        }
-
-        navigation.navigate({
-          name: "CloseVaultScreen",
-          params: {
-            vaultId: vault.vaultId,
-          },
-          merge: true,
-        });
-      },
-      testID: "vault_detail_close_vault",
-    },
-  ];
+    };
+  };
+  const {
+    bottomSheetRef,
+    containerRef,
+    dismissModal,
+    expandModal,
+    isModalDisplayed,
+    bottomSheetScreen,
+    setBottomSheetScreen,
+  } = useBottomSheet();
 
   useEffect(() => {
     const _vault = vaults.find((v) => v.vaultId === vaultId);
@@ -117,208 +150,257 @@ export function VaultDetailScreen({ route, navigation }: Props): JSX.Element {
     }
   }, [blockCount]);
 
+  useEffect(() => {
+    if (vault?.state === LoanVaultState.IN_LIQUIDATION) {
+      WalletAlert({
+        title: translate("components/VaultCard", "Liquidated vault"),
+        message: translate(
+          "screens/VaultDetailScreen",
+          "Vault has been liquidated, and now available for auction. View your vault on DeFiScan for more details."
+        ),
+        buttons: [
+          {
+            text: translate("screens/Settings", "Go back"),
+            style: "cancel",
+            onPress: async () => {
+              navigation.navigate("LoansScreen", {});
+            },
+          },
+          {
+            text: translate("screens/Settings", "View on Scan"),
+            style: "destructive",
+            onPress: async () => {
+              openURL(getVaultsUrl(vaultId));
+              navigation.navigate("LoansScreen", {});
+            },
+          },
+        ],
+      });
+    }
+  }, [vault?.state]);
+
+  const onAddPressed = () => {
+    if (vault === undefined) {
+      return;
+    }
+
+    setSnapPoints({
+      ios: ["75%"],
+      android: ["70%"],
+    });
+    setBottomSheetScreen([
+      {
+        stackScreenName: "TokenList",
+        component: BottomSheetTokenList({
+          tokenType: TokenType.CollateralItem,
+          vault: vault as LoanVaultActive,
+          onTokenPress: async (item) => {
+            dismissModal();
+            navigateToAddRemoveCollateralScreen(item, true);
+          },
+        }),
+        option: getBottomSheetHeader(
+          translate("screens/EditCollateralScreen", "Select Collateral")
+        ),
+      },
+    ]);
+    expandModal();
+  };
+
+  const navigateToAddRemoveCollateralScreen = (
+    collateralItem: CollateralItem | BottomSheetToken,
+    isAdd: boolean
+  ) => {
+    const addOrRemoveParams = {
+      vault: vault as LoanVaultActive,
+      collateralItem,
+      collateralTokens,
+      isAdd,
+    };
+
+    navigation.navigate({
+      name: "AddOrRemoveCollateralScreen",
+      params: addOrRemoveParams,
+    });
+  };
+
+  const onBorrowPressed = () => {
+    setSnapPoints({
+      ios: ["75%"],
+      android: ["70%"],
+    });
+    setBottomSheetScreen([
+      {
+        stackScreenName: "LoanTokensList",
+        component: BottomSheetLoanTokensList({
+          onPress: (item: LoanToken) => {
+            dismissModal();
+            navigation.navigate({
+              name: "BorrowLoanTokenScreen",
+              params: {
+                vault: vault,
+                loanToken: item,
+              },
+              merge: true,
+            });
+          },
+          loanTokens,
+          isLight,
+        }),
+        option: getBottomSheetHeader(
+          translate("components/BottomSheetLoanTokensList", "Select Token")
+        ),
+      },
+    ]);
+    expandModal();
+  };
+
+  const onPayPressed = () => {
+    if (vault === undefined || vault.state === LoanVaultState.IN_LIQUIDATION) {
+      return;
+    }
+
+    setSnapPoints({
+      ios: ["65%"],
+      android: ["60%"],
+    });
+    setBottomSheetScreen([
+      {
+        stackScreenName: "PayTokensList",
+        component: BottomSheetPayBackList({
+          onPress: (
+            item: LoanVaultTokenAmount,
+            isPayDUSDUsingCollateral: boolean
+          ) => {
+            dismissModal();
+            navigateToPayScreen(item, isPayDUSDUsingCollateral);
+          },
+          vault: vault,
+          data: vault.loanAmounts,
+          isLight: isLight,
+        }),
+        option: getBottomSheetHeader(
+          translate("screens/VaultDetailScreen", "Select Loan")
+        ),
+      },
+    ]);
+    expandModal();
+  };
+
+  const navigateToPayScreen = (
+    loanToken: LoanVaultTokenAmount,
+    isPayDUSDUsingCollateral: boolean
+  ) => {
+    navigation.navigate({
+      name: "PaybackLoanScreen",
+      merge: true,
+      params: {
+        vault: vault,
+        loanTokenAmount: loanToken,
+        isPaybackDUSDUsingCollateral: isPayDUSDUsingCollateral,
+      },
+    });
+  };
+
+  const onEditPressed = () => {
+    if (vault === undefined) {
+      return;
+    }
+
+    navigation.navigate({
+      name: "EditLoanSchemeScreen",
+      params: {
+        vaultId: vault.vaultId,
+      },
+      merge: true,
+    });
+  };
+
+  const onCloseVaultPressed = () => {
+    if (vault === undefined) {
+      return;
+    }
+
+    navigation.navigate({
+      name: "CloseVaultScreen",
+      params: {
+        vaultId: vault.vaultId,
+      },
+      merge: true,
+    });
+  };
+
   if (vault === undefined) {
     return <></>;
   }
 
   return (
-    <ThemedScrollView>
-      <ThemedView light={tailwind("bg-white")} dark={tailwind("bg-gray-800")}>
-        <View style={tailwind("p-4")}>
-          <VaultIdSection vault={vault} testID="vault_id_section" />
-          <VaultInfoSection vault={vault} />
-        </View>
-        <ThemedView
-          light={tailwind("border-gray-200")}
-          dark={tailwind("border-gray-700")}
-          style={tailwind("pb-4 border-b")}
-        >
-          <ScrollableButton
-            buttons={vaultActionButtons}
-            containerStyle={tailwind("pl-4")}
-          />
-        </ThemedView>
-      </ThemedView>
-      <VaultDetailTabSection vault={vault} tab={tab} />
-    </ThemedScrollView>
-  );
-}
+    <View ref={containerRef} style={tailwind("flex-1")}>
+      <ThemedScrollViewV2 contentContainerStyle={tailwind("w-full pb-12")}>
+        <VaultDetailStatus
+          vault={vault}
+          vaultStatus={vaultState?.status}
+          nextColRatio={nextCollateralizationRatio}
+        />
+        <VaultActionButtons
+          vaultStatus={vaultState?.status}
+          canUseOperations={canUseOperations}
+          onAddPressed={onAddPressed}
+          onBorrowPressed={onBorrowPressed}
+          onPayPressed={onPayPressed}
+          onEditPressed={onEditPressed}
+        />
+        <VaultDetailSummary
+          maxLoanAmount={maxLoanAmount}
+          totalCollateral={totalCollateralValue}
+          totalLoan={totalLoanAmount}
+          vaultId={vaultId}
+          interest={vault?.loanScheme?.interestRate}
+          minColRatio={vault?.loanScheme?.minColRatio}
+        />
+        <VaultDetailCollateralsRow
+          vault={vault}
+          collateralTokens={collateralTokens}
+          vaultStatus={vaultState?.status}
+          onAddPress={(collateralItem) => {
+            navigateToAddRemoveCollateralScreen(collateralItem, true);
+          }}
+          onRemovePress={(collateralItem) => {
+            navigateToAddRemoveCollateralScreen(collateralItem, false);
+          }}
+        />
+        <VaultDetailLoansRow onPay={navigateToPayScreen} vault={vault} />
+        <CloseVaultButton
+          vaultStatus={vaultState?.status}
+          canUseOperations={canUseOperations}
+          onCloseVaultPressed={onCloseVaultPressed}
+        />
 
-function VaultIdSection({
-  vault,
-  testID,
-}: {
-  vault: LoanVault;
-  testID: string;
-}): JSX.Element {
-  const { getVaultsUrl } = useDeFiScanContext();
-  const colRatio =
-    vault.state === LoanVaultState.IN_LIQUIDATION ? 0 : vault.informativeRatio;
-  const totalLoanAmount =
-    vault.state === LoanVaultState.IN_LIQUIDATION ? 0 : vault.loanValue;
-  const totalCollateralValue =
-    vault.state === LoanVaultState.IN_LIQUIDATION ? 0 : vault.collateralValue;
-  const vaultState = useVaultStatus(
-    vault.state,
-    new BigNumber(colRatio),
-    new BigNumber(vault.loanScheme.minColRatio),
-    new BigNumber(totalLoanAmount),
-    new BigNumber(totalCollateralValue)
-  );
-  const collateralAmounts =
-    vault.state === LoanVaultState.IN_LIQUIDATION
-      ? []
-      : vault.collateralAmounts;
-  const loanAmounts =
-    vault.state === LoanVaultState.IN_LIQUIDATION ? [] : vault.loanAmounts;
-  const nextCollateralizationRatio = useNextCollateralizationRatio(
-    collateralAmounts,
-    loanAmounts
-  );
-  return (
-    <>
-      <ThemedView
-        light={tailwind("bg-white")}
-        dark={tailwind("bg-gray-800")}
-        style={tailwind("flex flex-row items-center")}
-      >
-        <View style={tailwind("flex flex-1")}>
-          <View style={tailwind("flex flex-row mb-2 items-center")}>
-            <ThemedText
-              light={tailwind("text-gray-400")}
-              dark={tailwind("text-gray-500")}
-              style={tailwind("text-xs mr-1")}
-            >
-              {translate("screens/VaultDetailScreen", "Vault ID")}
-            </ThemedText>
-            <VaultStatusTag
-              status={vaultState.status}
-              testID="vault_detail_status"
-            />
-          </View>
-          <View style={tailwind("flex flex-row mb-2 items-center")}>
-            <ThemedText
-              testID="vault_detail_id"
-              style={tailwind("text-sm font-semibold w-8/12 flex-1 mr-2")}
-            >
-              {vault.vaultId}
-            </ThemedText>
-            <TouchableOpacity
-              onPress={async () =>
-                await openURL(getVaultsUrl(vault.vaultId ?? ""))
-              }
-            >
-              <ThemedIcon
-                dark={tailwind("text-darkprimary-500")}
-                iconType="MaterialIcons"
-                light={tailwind("text-primary-500")}
-                name="open-in-new"
-                size={22}
-                style={tailwind("-mr-1")}
-              />
-            </TouchableOpacity>
-          </View>
-        </View>
-      </ThemedView>
-      {vault.state !== LoanVaultState.IN_LIQUIDATION &&
-        vaultState.status !== VaultStatus.Empty &&
-        vaultState.status !== VaultStatus.Ready && (
-          <CollateralizationRatioDisplay
-            collateralizationRatio={vault.informativeRatio}
-            minCollateralizationRatio={vault.loanScheme.minColRatio}
-            totalLoanAmount={vault.loanValue}
-            nextCollateralizationRatio={nextCollateralizationRatio?.toFixed(8)}
-            testID={testID}
+        {Platform.OS === "web" ? (
+          <BottomSheetWebWithNavV2
+            modalRef={containerRef}
+            screenList={bottomSheetScreen}
+            isModalDisplayed={isModalDisplayed}
+            // eslint-disable-next-line react-native/no-inline-styles
+            modalStyle={{
+              position: "absolute",
+              bottom: "0",
+              height: "404px",
+              width: "375px",
+              zIndex: 50,
+              borderTopLeftRadius: 15,
+              borderTopRightRadius: 15,
+              overflow: "hidden",
+            }}
+          />
+        ) : (
+          <BottomSheetWithNavV2
+            modalRef={bottomSheetRef}
+            screenList={bottomSheetScreen}
+            snapPoints={snapPoints}
           />
         )}
-    </>
-  );
-}
-
-function VaultInfoSection(props: { vault?: LoanVault }): JSX.Element | null {
-  if (props.vault === undefined) {
-    return null;
-  }
-
-  if (props.vault.state === LoanVaultState.IN_LIQUIDATION) {
-    return (
-      <View style={tailwind("flex -mb-2")}>
-        <VaultSectionTextRow
-          value={props.vault.batchCount}
-          lhs={translate("screens/VaultDetailScreen", "Auction batches")}
-          testID="text_auction_batches"
-        />
-      </View>
-    );
-  }
-
-  return (
-    <View style={tailwind("flex -mb-2")}>
-      {props.vault.state === LoanVaultState.ACTIVE &&
-      props.vault.collateralValue === "0" &&
-      props.vault.loanValue === "0" ? (
-        <>
-          <VaultSectionTextRow
-            value={props.vault.loanScheme.minColRatio}
-            lhs={translate(
-              "screens/VaultDetailScreen",
-              "Min. collateralization ratio"
-            )}
-            testID="text_min_col_ratio"
-            suffixType="text"
-            suffix="%"
-            info={{
-              title: "Min. collateralization ratio",
-              message:
-                "Minimum required collateralization ratio based on loan scheme selected. A vault will go into liquidation when the collateralization ratio goes below the minimum requirement.",
-            }}
-          />
-          <VaultSectionTextRow
-            value={props.vault.loanScheme.interestRate}
-            lhs={translate("screens/VaultDetailScreen", "Vault interest (APR)")}
-            testID="text_vault_interest"
-            suffixType="text"
-            suffix="%"
-            info={{
-              title: "Annual vault interest",
-              message:
-                "Annual vault interest rate based on the loan scheme selected.",
-            }}
-          />
-        </>
-      ) : (
-        <>
-          <VaultSectionTextRow
-            value={getPrecisedTokenValue(props.vault.collateralValue)}
-            lhs={translate(
-              "screens/VaultDetailScreen",
-              "Total collateral (USD)"
-            )}
-            testID="text_total_collateral_value"
-            prefix="$"
-            isOraclePrice
-          />
-          <VaultSectionTextRow
-            value={getPrecisedTokenValue(props.vault.loanValue)}
-            lhs={translate("screens/VaultDetailScreen", "Total loans (USD)")}
-            testID="text_total_loan_value"
-            prefix="$"
-            isOraclePrice
-          />
-          <VaultSectionTextRow
-            value={props.vault.loanScheme.interestRate}
-            lhs={translate("screens/VaultDetailScreen", "Vault interest (APR)")}
-            testID="text_vault_interest"
-            suffixType="text"
-            suffix="%"
-            info={{
-              title: "Annual vault interest",
-              message:
-                "Annual vault interest rate based on the loan scheme selected.",
-            }}
-          />
-        </>
-      )}
+      </ThemedScrollViewV2>
     </View>
   );
 }
